requirements.txt
Django
djangorestframework
django-filter
pytest
pytest-django
psycopg2-binary
requests
requests_mock
python-dotenv
dj-database-url
isort
ruff
black